Issue #SC-211 feat:Change as per review
