+++ service/src/main/resources/application.conf
@@ -226,13 +226,13 @@ SunbirdMWSystem {
             nr-of-instances = 2
             dispatcher = rr-usr-dispatcher
           }
-        "/BackgroundRequestRouter/*/DataSecurityBackgroundActor"
+        "/BackgroundRequestRouter/*/BackgroundUserDataEncryptionActor"
           {
             router = smallest-mailbox-pool
-            nr-of-instances = 2
+            nr-of-instances = 4
             dispatcher = rr-usr-dispatcher
           }
-        "/RequestRouter/*/DataSecurityActor"
+        "/RequestRouter/*/UserDataEncryptionActor"
           {
             router = smallest-mailbox-pool
             nr-of-instances = 2
